Update and properly isolate build environment for VS 2019.
This fixes a number of issues that prevented the dependency build
environment from functioning as intended with Visual Studio 2019.
Additionally, the output of this build system is, on Windows, sufficient
to build blender without depending on the legacy bf-blender subversion.

NOTE: The following must be true when running build_deps.cmd.
- Python must be on the PATH.
- Must be run from the Visual Studio Developer's prompt.
- Must be run with administrator permissions such that mklink can be
used.